app/src/main/java/com/example/tasklist_movil/RegisterActivity.kt
package com.example.tasklist_movil

import android.content.Intent
import android.os.Bundle
import android.widget.Toast
import androidx.appcompat.app.AppCompatActivity
import com.example.tasklist_movil.databinding.ActivityRegisterBinding
import com.google.firebase.auth.FirebaseAuth

class RegisterActivity : AppCompatActivity() {

    lateinit var binding: ActivityRegisterBinding
    lateinit var firebaseAuth: FirebaseAuth

    private val messageLog = mutableListOf<String>()
    fun logMessage(message: String) {
        messageLog.add(message)
    }

    override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        binding = ActivityRegisterBinding.inflate(layoutInflater)
        setContentView(binding.root)

        firebaseAuth = FirebaseAuth.getInstance()

        binding.tvGoToLogin.setOnClickListener {
            goToLogin()
        }

        binding.Registrarsebtn.setOnClickListener {
            val name = binding.NameEditText.text.toString()
            val email = binding.EmailEditText.text.toString()
            val pass = binding.PasswordEditText.text.toString()
            val confirmPass = binding.ConfirPasswordEditText.text.toString()

            registerUser(name, email, pass, confirmPass)
        }
    }

    fun registerUser(name: String, email: String, pass: String, confirmPass: String) {
        if (name.isNotEmpty() && email.isNotEmpty() && pass.isNotEmpty() && confirmPass.isNotEmpty()) {
            if (pass == confirmPass) {
                firebaseAuth.createUserWithEmailAndPassword(email, pass).addOnCompleteListener { task ->
                    if (task.isSuccessful) {
                        val intent = Intent(this, Login::class.java)
                        startActivity(intent)
                    } else {
                        Toast.makeText(this, "Error al registrar el usuario: ${task.exception}", Toast.LENGTH_SHORT).show()
                    }
                }
            } else {
                Toast.makeText(this, "Las contraseñas no coinciden", Toast.LENGTH_SHORT).show()
            }
        } else {
            Toast.makeText(this, "Existen campos vacíos", Toast.LENGTH_SHORT).show()
        }
    }

    private fun goToLogin() {
        val intent = Intent(this, Login::class.java)
        startActivity(intent)
    }


}